pyproject.toml
[project]
name = "youtube-smart-pause"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
